+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/BaseKotlinExampleActivity.kt
@@ -3,15 +3,21 @@ package com.mapbox.search.sample.api
 import android.annotation.SuppressLint
 import android.os.Bundle
 import android.util.Log
+import android.view.View
+import android.widget.ScrollView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.AppCompatButton
 import androidx.appcompat.widget.Toolbar
 import com.google.gson.GsonBuilder
+import com.mapbox.annotation.MapboxExperimental
+import com.mapbox.search.offline.OfflineSearchResult
 import com.mapbox.search.sample.R
 
+@OptIn(MapboxExperimental::class)
 abstract class BaseKotlinExampleActivity : AppCompatActivity() {
 
+    private val logTextScroll: ScrollView by lazy { findViewById(R.id.logTextScroll) }
     private val toolbarView: Toolbar by lazy { findViewById(R.id.toolbar) }
     private val logTextView: TextView by lazy { findViewById(R.id.logText) }
     private val startButton: AppCompatButton by lazy { findViewById(R.id.startButton) }
@@ -42,28 +48,62 @@ abstract class BaseKotlinExampleActivity : AppCompatActivity() {
         startButton.isEnabled = true
     }
 
-    protected fun logI(tag: String, message: String, result: Any? = null) {
-        logTextView.post {
-            Log.i(tag, message.addToLog())
-            result?.let { res -> printObjectToLog(res) }
+    // TODO we probably shouldn't try to print in JSON format by default,
+    // because it prints all the object structure, including internal stuff
+    protected fun logI(tag: String = LOG_TAG, message: String, result: Any? = null) {
+        runOnUiThread {
+            Log.i(tag, message)
+            appendToLogMsg(message)
+            result?.let { res ->
+                Log.i(tag, message)
+                appendToLogMsg(prettify(res))
+            }
         }
     }
 
-    protected fun logE(tag: String, message: String, error: Throwable? = null) {
-        Log.i(tag, "$message: ${error?.message}".addToLog(), error)
+    protected fun logE(tag: String = LOG_TAG, message: String, error: Throwable? = null) {
+        runOnUiThread {
+            Log.e(tag, message, error)
+            appendToLogMsg("$message: ${error?.message}")
+        }
     }
 
-    protected fun prettify(result: Any): String {
-        val gson = GsonBuilder().setPrettyPrinting().serializeNulls().create()
-        return gson.toJson(result).toString()
-    }
-
-    private fun printObjectToLog(results: Any) {
-        prettify(results).addToLog()
+    protected fun printMessage(message: String) {
+        runOnUiThread {
+            Log.i(LOG_TAG, message)
+            appendToLogMsg(message)
+        }
     }
 
     @SuppressLint("SetTextI18n")
-    protected fun String.addToLog() = this.also {
-        logTextView.text = "${logTextView.text}\n$this"
+    private fun appendToLogMsg(message: String) {
+        logTextView.text = "${logTextView.text}\n$message"
+        logTextScroll.post {
+            logTextScroll.fullScroll(View.FOCUS_DOWN)
+        }
+    }
+
+    protected companion object {
+
+        const val LOG_TAG = "SearchApiExample"
+
+        private val GSON = GsonBuilder().setPrettyPrinting().serializeNulls().create()
+
+        fun prettify(result: Any) = GSON.toJson(result).toString()
+
+        fun OfflineSearchResult.toPrettyString(): String {
+            return "OfflineSearchResult(\n" +
+                    "\tid='$id',\n" +
+                    "\tmapboxId='$mapboxId',\n" +
+                    "\tname='$name',\n" +
+                    "\tdescriptionText=$descriptionText,\n" +
+                    "\taddress=$address,\n" +
+                    "\tcoordinate=$coordinate,\n" +
+                    "\troutablePoints=$routablePoints,\n" +
+                    "\tnewType=$newType,\n" +
+                    "\tdistanceMeters=$distanceMeters,\n" +
+                    "\tmetadata=$metadata\n" +
+                    ")"
+        }
     }
 }